+++ tt_eager/tt_dnn/op_library/moreh_groupnorm_backward/input_grad/kernels/dataflow/reader_moreh_groupnorm_backward_input_grad_large.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_eager/tt_dnn/op_library/moreh_groupnorm/kernel_utils/common.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/moreh_common.hpp"
 
 void kernel_main() {
     int i{0};
