+++ tt_eager/tt_dnn/op_library/moreh_layernorm/kernels/reader_moreh_layernorm_small.cpp
@@ -2,10 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <stdint.h>
-
-#include "dataflow_api.h"
-#include "tt_eager/tt_dnn/op_library/moreh_layernorm/kernels/utils.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/moreh_common.hpp"
 
 void kernel_main() {
     const auto input_addr = get_arg_val<uint32_t>(0);
