+++ tt_eager/tt_dnn/op_library/moreh_layernorm_backward/kernels/reader_moreh_layernorm_backward_gamma_beta_grad.cpp
@@ -2,10 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <stdint.h>
-
-#include "dataflow_api.h"
-#include "tt_eager/tt_dnn/op_library/moreh_layernorm_backward/kernels/utils.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/moreh_common.hpp"
 
 void kernel_main() {
     const auto output_grad_addr = get_arg_val<uint32_t>(0);
