--- tt_eager/tt_dnn/op_library/moreh_nll_loss_backward/kernel_utils/common.hpp
@@ -1,80 +1,0 @@
-// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#include <stdint.h>
-
-#include "dataflow_api.h"
-
-void fill_cb_with_value(uint32_t cb_id, uint32_t value) {
-    cb_reserve_back(cb_id, 1);
-    auto ptr = reinterpret_cast<uint16_t *>(get_write_ptr(cb_id));
-    for (int j = 0; j < 1024; j++) {
-        ptr[j] = uint16_t(value >> 16);
-    }
-    cb_push_back(cb_id, 1);
-}
-
-void mask_tile_hw(uint32_t l1_addr, uint32_t mask_h = 32, uint32_t mask_w = 32) {
-    union {
-        float f;
-        uint32_t u;
-    } zero;
-    zero.f = 0.0f;
-    const auto u16_zero = uint16_t(zero.u >> 16);
-
-    volatile tt_l1_ptr uint16_t* ptr = reinterpret_cast<volatile tt_l1_ptr uint16_t*>(l1_addr);
-    for (uint32_t h = 0; h < 16; h++) {
-        // sub tile 0
-        {
-            uint32_t mask_w_0 = (mask_w >= 16) ? 16 : mask_w;
-            uint32_t mask_h_0 = (mask_h >= 16) ? 16 : mask_h;
-            uint32_t w = (h >= mask_h_0) ? 0 : mask_w_0;
-            for (; w < 16; w++) {
-                ptr[h * 16 + w] = uint16_t(zero.u >> 16);
-            }
-        }
-        // sub tile 1
-        {
-            uint32_t mask_w_1 = (mask_w < 16) ? 0 : mask_w - 16;
-            uint32_t mask_h_0 = (mask_h >= 16) ? 16 : mask_h;
-            uint32_t w = (h >= mask_h_0) ? 0 : mask_w_1;
-            for (; w < 16; w++) {
-                ptr[h * 16 + w + 256] = uint16_t(zero.u >> 16);
-            }
-        }
-        // sub tile 2
-        {
-            uint32_t mask_w_0 = (mask_w >= 16) ? 16 : mask_w;
-            uint32_t mask_h_1 = (mask_h < 16) ? 0 : mask_h - 16;
-            uint32_t w = (h >= mask_h_1) ? 0 : mask_w_0;
-            for (; w < 16; w++) {
-                ptr[h * 16 + w + 512] = uint16_t(zero.u >> 16);
-            }
-        }
-        // sub tile 3
-        {
-            uint32_t mask_w_1 = (mask_w < 16) ? 0 : mask_w - 16;
-            uint32_t mask_h_1 = (mask_h < 16) ? 0 : mask_h - 16;
-            uint32_t w = (h >= mask_h_1) ? 0 : mask_w_1;
-            for (; w < 16; w++) {
-                ptr[h * 16 + w + 768] = uint16_t(zero.u >> 16);
-            }
-        }
-    }
-}
-
-void mask_tile_if_need(uint32_t l1_addr, uint32_t origin_h, uint32_t origin_w) {
-    constexpr uint32_t TILE_H = 32;
-    constexpr uint32_t TILE_W = 32;
-
-    const bool do_mask_h = (origin_h % TILE_H) != 0;
-    const uint32_t mask_h = do_mask_h ? (origin_h % TILE_H) : TILE_H;
-
-    const bool do_mask_w = (origin_w % TILE_W) != 0;
-    const uint32_t mask_w = do_mask_w ? (origin_w % TILE_W) : TILE_W;
-
-    if (do_mask_h || do_mask_w) {
-        mask_tile_hw(l1_addr, mask_h, mask_w);
-    }
-}
