+++ tt_eager/tt_dnn/op_library/moreh_norm/moreh_norm_h/kernels/reader_moreh_norm_h.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_eager/tt_dnn/op_library/moreh_norm/kernel_utils/common.hpp"
+#include "tt_eager/tt_dnn/kernels/dataflow/moreh_common.hpp"
 
 void kernel_main() {
     int i{0};
