+++ tt_eager/tt_dnn/op_library/moreh_sgd/kernels/writer_moreh_sgd.cpp
@@ -3,8 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "dataflow_api.h"
-#include "tt_eager/tt_dnn/op_library/moreh_sgd/kernels/common.hpp"
-#include "debug/dprint.h"
+#include "tt_eager/tt_dnn/kernels/dataflow/moreh_common.hpp"
 
 void kernel_main() {
     uint32_t i = 0;
